Send left stripe to the left neighbor. Dirty code.
Each processor sends its left stripe to the left neighbor. The neighbor
receives the stripe and saves it to the right column of the mGlobalPixels
array.

Next: refactor the code.
i) introduce a strategy for handling both periodic and non-periodic BC in
the horizontal direction (in the local WUF and in the
mergeLabelsAcrossProcessors);
ii) make cleaner code: put sending/receiving stripes, even/odd checking in
separate functions.