Both left and right stripes are sent.
Both left and right stripes of pixels are sent to and received from the
neighboring processors.
The code is a bit dirty and hard coded.

Next step: use the mGlobalPixels to initialize the global union-find.
Then run this global UF and record the merges that occur for the global
labels.