Started merging global labels.
Started the final phase of the Stage 3 of the algorithm.

As there are no cells in the current implementation (in contrast to the paper) a slightly
different approach is used.

At first, the extended (i.e. with the stripes from the neighboring
processors) pixels are merged based on their global cluster property. In
this way we form a local UF that contains the roots of the clusters. After
that we again run the UF algorithm, but now we merge the pixels based on
their value. If the merge occurs, we record the global ids of the clusters
that have been merged.