Local UF of the global cluster ids.
Works: running the local union-find on the extended pixels. This gives us
the initialization of the roots based on the local pixel ids. Merging
occurs based on the global cluster id and not on the pixel values, which
allows us to reproduce the clusters identified on the Stage 2.

After this phase run the local UF again on the same data, but merge based
on the pixel value and record merges.